chore(analyser/go): separate util components
